Fix unicode domain length limit
Punycode's `encode` function does not suit our purposes. It does not prefix the encoded string with `xn--`, which is an inherent part of punycoded domains, and adds an extra `-` to the end when the domain is already punycoded. We switch these calls over to `toASCII`, which performs as expected in those two areas. Moreover, we were already using `toASCII` in `canonicalizeAtom`, so this makes our use of the punycode library consistent.